requirements.txt
appdirs==1.4.3
distlib==0.3.0
filelock==3.0.12
Jinja2==2.11.1
MarkupSafe==1.1.1
PyYAML==5.3
six==1.14.0
virtualenv==20.0.3
